refactor: add convertUtcNumberToOffset on the convert-date file
